解决无法记录带小数点的工作量信息

+++ src/main/java/com/raymon/workrecord/entity/WorkLoad.java
@@ -23,7 +23,7 @@ public class WorkLoad {
 	
 	// 工作时长
 	@TableField("work_time")
-	private Integer workTime;
+	private Double workTime;
 	
 	@TableField("user_id")
 	private String userId;
